Moving the services to Webjobs Script project
