+++ src/WebJobs.Script.WebHost/Middleware/IJobHostHttpMiddleware.cs
